+++ CHANGES.md
@@ -7,6 +7,7 @@ Changes in 10.7
 - JS add Select2 jQuery plugin in assets/js/jquery-select2/
 - Add Select2Input() function in Inputs.php
 - Use Select2 instead of Chosen, fixes the overflow issue in Widget.php, MakeReferral.php, Transcripts.fnc.php & Address.inc.php
+- Fix PHP8.2 deprecated Creation of dynamic property in Markdownify/ConverterExtra.php
 
 Changes in 10.6.3
 -----------------